# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lojistik Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loads:
    get:
      operationId: listLoads
      tags: [loads]
      summary: List load listings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, pending, assigned, completed, cancelled]
        - name: vehicleType
          in: query
          schema:
            type: string
        - name: loadType
          in: query
          schema:
            type: string
        - name: origin
          in: query
          schema:
            type: string
        - name: destination
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of loads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoadListResponse"
    post:
      operationId: createLoad
      tags: [loads]
      summary: Create load listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoadRequest"
      responses:
        "201":
          description: Created load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
  /loads/{id}:
    get:
      operationId: getLoad
      tags: [loads]
      summary: Get load details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
    patch:
      operationId: updateLoad
      tags: [loads]
      summary: Update load
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLoadRequest"
      responses:
        "200":
          description: Updated load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List offers
      parameters:
        - name: loadId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, withdrawn]
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferListResponse"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Submit offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferRequest"
      responses:
        "201":
          description: Created offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}/accept:
    post:
      operationId: acceptOffer
      tags: [offers]
      summary: Accept offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}/reject:
    post:
      operationId: rejectOffer
      tags: [offers]
      summary: Reject offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offer rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pickup, in_transit, delivered, cancelled]
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentListResponse"
  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get shipment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, corporate, individual, driver]
        status:
          type: string
          enum: [active, suspended, pending]
        company:
          type: string
        avatarUrl:
          type: string
        rating:
          type: number
        totalShipments:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - status
        - createdAt
...
    Load:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        origin:
          type: string
        destination:
          type: string
        distance:
          type: number
        weight:
          type: number
        loadType:
          type: string
        vehicleType:
          type: string
        pricingModel:
          type: string
          enum: [fixed, bidding]
        price:
          type: number
        minBid:
          type: number
        maxBid:
          type: number
        status:
          type: string
          enum: [active, pending, assigned, completed, cancelled]
        isPremium:
          type: boolean
        postedBy:
          $ref: "#/components/schemas/User"
        offersCount:
          type: integer
        pickupDate:
          type: string
          format: date-time
        deliveryDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        originLat:
          type: number
        originLng:
          type: number
        destLat:
          type: number
        destLng:
          type: number
      required:
        - id
        - title
        - origin
        - destination
        - loadType
        - vehicleType
        - pricingModel
        - status
        - createdAt
    LoadListResponse:
      type: object
      properties:
        loads:
          type: array
          items:
            $ref: "#/components/schemas/Load"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - loads
        - total
        - page
        - pageSize
    CreateLoadRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        origin:
          type: string
        destination:
          type: string
        weight:
          type: number
        loadType:
          type: string
        vehicleType:
          type: string
        pricingModel:
          type: string
          enum: [fixed, bidding]
        price:
          type: number
        minBid:
          type: number
        pickupDate:
          type: string
          format: date-time
        deliveryDate:
          type: string
          format: date-time
      required:
        - title
        - origin
        - destination
        - loadType
        - vehicleType
        - pricingModel
    UpdateLoadRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, pending, assigned, completed, cancelled]
        price:
          type: number
        isPremium:
          type: boolean
    Offer:
      type: object
      properties:
        id:
          type: string
        loadId:
          type: string
        load:
          $ref: "#/components/schemas/Load"
        driver:
          $ref: "#/components/schemas/User"
        amount:
          type: number
        note:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - loadId
        - amount
        - status
        - createdAt
    OfferListResponse:
      type: object
      properties:
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        total:
          type: integer
      required:
        - offers
        - total
    CreateOfferRequest:
      type: object
      properties:
        loadId:
          type: string
        amount:
          type: number
        note:
          type: string
      required:
        - loadId
        - amount
    Shipment:
      type: object
      properties:
        id:
          type: string
        loadId:
          type: string
        load:
          $ref: "#/components/schemas/Load"
        driver:
          $ref: "#/components/schemas/User"
        status:
          type: string
          enum: [pickup, in_transit, delivered, cancelled]
        currentLat:
          type: number
        currentLng:
          type: number
        estimatedArrival:
          type: string
          format: date-time
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentEvent"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - loadId
        - status
        - createdAt
    ShipmentEvent:
      type: object
      properties:
        id:
          type: string
        event:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
        lat:
          type: number
        lng:
          type: number
      required:
        - id
        - event
        - timestamp
    ShipmentListResponse:
      type: object
      properties:
        shipments:
          type: array
          items:
            $ref: "#/components/schemas/Shipment"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - shipments
        - total
        - page
        - pageSize
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [new_load, offer_received, offer_accepted, offer_rejected, shipment_update, system]
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        relatedId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - body
        - isRead
        - createdAt
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required:
        - notifications
        - unreadCount
...