# lib/api-zod/src/generated/types/createOfferRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lojistik Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOfferRequest {
  loadId: string;
  amount: number;
  note?: string;
}