# lib/api-zod/src/generated/types/loadListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lojistik Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Load } from "./load";

export interface LoadListResponse {
  loads: Load[];
  total: number;
  page: number;
  pageSize: number;
}