# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lojistik Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: string;
  name: string;
  email: string;
  phone?: string;
  role: UserRole;
  status: UserStatus;
  company?: string;
  avatarUrl?: string;
  rating?: number;
  totalShipments?: number;
  createdAt: Date;
}